+++ packages/express/src/index.ts
@@ -59,6 +59,10 @@ export const apilyticsMiddleware = (
         requestSize,
         responseSize,
         userAgent: req.headers['user-agent'],
+        // Type assertion since the value cannot be an array.
+        ip: (req.headers['x-forwarded-for'] as string | undefined)
+          ?.split(',')[0]
+          .trim(),
         timeMillis: timer(),
         apilyticsIntegration: 'apilytics-node-express',
         integratedLibrary: EXPRESS_VERSION